# openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Skillex Africa API",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://{{base_url}}"
		},
		{
			"url": "http://{{base_url_live}}"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
...
	"paths": {
...
		"/admin/users": {
			"get": {
				"tags": ["admin > users"],
				"summary": "get_single_user",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "userId",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {}
						}
					}
				}
			}
		},
...
	}
}